refactor(elm-2): move the settings-reset to a mode-agnostic home
It now dispatches over every mode's pick table, so living under services/pickem/
made the first mode to ship the owner of a rule that is not its own — the
squatting the mode-naming rule exists to prevent, and the next reader looking
for Survivor's reset would not have found it there.

services/leagues/ is where it belongs: the reset runs inside updateLeague's
transaction and is about league settings, not about either mode's picks. It
stays out of the leagues barrel for the same reason serialize.ts does — crud.ts
is its only caller.
