+++ apps/api/src/services/leagues/settings-reset.ts
